Logo and buttons two versions
I made two versions of the logo and some buttons. One is a higher definition version. They should all be transparent and are intended to be put on a black screen (helps with readability) The buttons are very small bc I was too lazy to scale them up because I believe you can do that in unity. If not I can submit scaled up versions. I didn't know exactly what buttons we need, so I just made what I could think of